% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_accumulate_band_ic.R
\name{accumulate_img_bands_to_list}
\alias{accumulate_img_bands_to_list}
\title{accumulate bands to list}
\usage{
accumulate_img_bands_to_list(img, lst)
}
\arguments{
\item{img}{ee Image class object}

\item{lst}{list containing first img to iterate over}
}
\value{
ee List of images with each item containing the successive accumulation of
}
\description{
Function to be used with [rgee] `iterate`  to return a list of images containing the image
}
