% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_mask_modis.R
\name{bitwise_extract}
\alias{bitwise_extract}
\title{bitwise_extract}
\usage{
bitwise_extract(value, from_bit, to_bit)
}
\arguments{
\item{value}{ee$Number or ee$Image to extract from}

\item{from_bit}{int or ee$Number with the first bit}

\item{to_bit}{int or ee$number with the last bit (inclusive)}
}
\value{
bit
}
\description{
utility to extract bitmask values - https://spatialthoughts.com/2021/08/19/qa-bands-bitmasks-gee/
}
