% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_wrangling_functions.R
\name{clean_numeric_long}
\alias{clean_numeric_long}
\title{Title}
\usage{
clean_numeric_long(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Title
}
\examples{

col_climate_risk.riesgo_cc numeric or cat
}
