% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_mask_modis.R
\name{cloud_and_quality_mask_modis}
\alias{cloud_and_quality_mask_modis}
\title{cloud_and_quality_mask_modis}
\usage{
cloud_and_quality_mask_modis(x, ...)
}
\arguments{
\item{x}{image or image collection}

\item{...}{Other arguments to pass to}
}
\value{
image/image collection masked
}
\description{
cloud_and_quality_mask_modis
}
\examples{
\dontrun{
library(rgee)
library(easyrgee)
ee_Initialize()

modis_ic = ee$ImageCollection('MODIS/006/MYD13Q1')$
    filter(ee$Filter$date('2015-08-01', '2015-08-31'))

modis_first_masked<- cloud_and_quality_mask_modis(modis_ic$first())

# just demonstrating that it also works on ic
modis_ic_masked <-  cloud_and_quality_mask_modis(modis_ic)

# now we can visualize
# lets make a point to zoom in on
ptdf <- data.frame(lon=44.43604342660201,lat=3.39908380146876)
ptsf <- sf::st_as_sf(ptdf, coords=c("lon","lat"),crs=4326)
pt_ee <- sf_as_ee(ptsf)

# center and zoom
Map$centerObject(pt_ee, 13)

# here is a standard MODIS ndvi viz set up
ndviVis = list(
    min=0.0,
    max= 8000.0,
    palette= c(
        'FFFFFF', 'CE7E45', 'DF923D', 'F1B555', 'FCD163', '99B718', '74A901',
        '66A000', '529400', '3E8601', '207401', '056201', '004C00', '023B01',
        '012E01', '011D01', '011301'
        )
)

Map$addLayer(modist_first_masked$select("NDVI"),ndviVis)


}
}
