% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_mask_modis.R
\name{cloud_mask_modis}
\alias{cloud_mask_modis}
\alias{cloud_mask_modis.ee.image.Image}
\alias{cloud_mask_modis.ee.imagecollection.ImageCollection}
\title{mask_modis_clouds}
\usage{
cloud_mask_modis(x, ...)

\method{cloud_mask_modis}{ee.image.Image}(x)

\method{cloud_mask_modis}{ee.imagecollection.ImageCollection}(x)
}
\arguments{
\item{x}{MODIS `ee$Image` or `ee$ImageCollection`}
}
\description{
modis_mask_clouds : function to mask clouds, cloud shadows, snow.
}
\examples{
\dontrun{
 library(easyrgee)
 library(rgee)
 ee_Initialize()
 modisIC <- ee$ImageCollection("MODIS/006/MOD13Q1")
 # mask an imageCollection
 modisIC_cloud_masked <- cloud_mask_modis(modisIC)
 # mask an image
 modis_first_cloud_masked <- cloud_mask_modis(modisIC$first())

}
}
