% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_air_quality_idicators.R
\name{cloud_mask_s5p}
\alias{cloud_mask_s5p}
\title{Title}
\usage{
cloud_mask_s5p(image, threshold)
}
\arguments{
\item{image}{image from sentinel 5p sensor}

\item{threshold}{cloud fraction threshold (default= 0.05)}
}
\value{
`ee$Image` with pixels with cloud fraction less than threshold masked
}
\description{
Title
}
