% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_wrangling_functions.R
\name{compile_clean_rs_categoricals}
\alias{compile_clean_rs_categoricals}
\title{Title}
\usage{
compile_clean_rs_categoricals(country_code, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
                                             col_chirps_rainfall_intensity,
                                             col_prev_3mo_drought_modis_basea,
                                             col_chirps_spi,
                                             col_dist_to_coast,
                                             col_landforms_reclassified,
                                             col_oxford_access,
                                             col_npp,
                                             col_local_value_merged)
}
}
