% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_wrangling_functions.R
\name{compile_rs_categorical_long}
\alias{compile_rs_categorical_long}
\title{Title}
\usage{
compile_rs_categorical_long(country_code, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

}
