% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouple_coordinates.R
\name{decouple_coordinates2}
\alias{decouple_coordinates2}
\title{setup geo data}
\usage{
decouple_coordinates2(
  df,
  uuid,
  lon,
  lat,
  cols_to_keep = NULL,
  country_code = NULL,
  pw
)
}
\arguments{
\item{df}{data.frame}

\item{uuid}{\code{character} uuid column name}

\item{lon}{\code{numeric} longitude column name}

\item{lat}{\code{numeric} latitude column name}
}
\value{
a list containing a simple feature spatial object with a new unique ID (geo data)
}
\description{
set up survey data for RS extraction
}
\examples{
\dontrun{
dat<- read_csv(here::here("dat/assessment_data.csv"))
setup_geo_data(df = dat,
              uuid = "_uuid",
              lon = "_gps_reading_longitude",
              lat = "_gps_reading_latitude")
}
}
