% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_accum_reduce_to_doy.R
\name{ee_accum_reduce_to_doy}
\alias{ee_accum_reduce_to_doy}
\title{creates day of year image/raster based on an accumulation threshold}
\usage{
ee_accum_reduce_to_doy(ic, thresh, band)
}
\arguments{
\item{ic}{an ee ImageCollection}

\item{thresh}{accumulation threshold}

\item{band}{band to accumulate}
}
\value{
and ee image/raster the contains the first date (DOY) the accumulation threshold was met per pixel.
}
\description{
creates day of year image/raster based on an accumulation threshold
}
\details{
user should carefully decide start and end date for accumulation. This is currently really only designed for assessing seasonal accumulations
}
\examples{
\dontrun{

# load libraries
library(easyrgee)
library(rgee)
ee_Initialize()

chirps <-  ee$ImageCollection("UCSB-CHG/CHIRPS/DAILY")$filterDate("2016-01-01","2016-12-31")
precip_cumulative<- ee_accumulate_band_ic(ic = chirps, band = "precipitation")
#visualize
Map$addLayer(precip_threshold_doy,
             visParams = list(min=0,
                           max=365,
                             palette=c('#9ecae1,
                             #ffffff,
                              #ffeda0,
                              #feb24c,
                              #f03b20'))
             )

}
}
