% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_accumulate_band_ic.R
\name{ee_accumulate_band_ic}
\alias{ee_accumulate_band_ic}
\title{accumulate_ic_band}
\usage{
ee_accumulate_band_ic(ic, band)
}
\arguments{
\item{ic}{`ee$ImageCollection` object with user-defined start and end-date}

\item{band}{band to accumulate value from}
}
\value{
returns `ee$ImageCollection` with each contained image containing the successive addition/accumulation of pixel-level values of chosen band.
}
\description{
Creates a cumulative precipitation [rgee] `ee$ImageCollection` with each all pixels in each consecutive image containing the sum of all previous values
}
\details{
user should carefully decide start and end date for accumulation. This is currently really only designed for assessing seasonal accumulations
}
\examples{
\dontrun{

# load libraries
library(easyrgee)
library(rgee)
ee_Initialize()

chirps <-  ee$ImageCollection("UCSB-CHG/CHIRPS/DAILY")$filterDate("2016-01-01","2016-12-31")
precip_cumulative<- ee_accumulate_band_ic(ic = chirps, band = "precipitation")

}
}
