% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.R
\name{ee_chirps_spi}
\alias{ee_chirps_spi}
\title{ee_chirps_spi calculate standard precipitation index from chirps daily}
\usage{
ee_chirps_spi(
  x = NULL,
  window,
  window_unit,
  moi,
  .load_chirps = T,
  band = "precipitation"
)
}
\arguments{
\item{x}{tidyee ee$ImageCollection}

\item{window}{window to calculate spi over}

\item{window_unit}{time unit of window (month only)}

\item{moi}{month of interest. (i.e if moi =4 and window is 3 months then the SPI is FMA)}

\item{band}{name of band}
}
\value{
tidyee object containing SPI ee$ImageCollection and vrt
}
\description{
ee_chirps_spi calculate standard precipitation index from chirps daily
}
\examples{
\dontrun{
library(rgee)
library(tidyrgee)
ee_Initialize()

spi_2_chirps<- ee_chirps_spi(x = chirps_tidy,window = 2,window_unit = "month")
}
}
