% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_closest_distance_to_val.R
\name{ee_closest_distance_to_val}
\alias{ee_closest_distance_to_val}
\alias{ee_closest_distance_to_val.ee.imagecollection.ImageCollection}
\alias{ee_closest_distance_to_val.ee.image.Image}
\title{ee_closest_distance_to_val}
\usage{
ee_closest_distance_to_val(
  x,
  y,
  boolean = "=",
  val = 2,
  scale,
  tidy_extract = T,
  via
)

\method{ee_closest_distance_to_val}{ee.imagecollection.ImageCollection}(
  x,
  y,
  boolean = "=",
  val = 2,
  scale,
  tidy_extract = T,
  via
)

\method{ee_closest_distance_to_val}{ee.image.Image}(
  x,
  y,
  boolean = "=",
  val = 2,
  scale,
  tidy_extract = T,
  via = via
)
}
\arguments{
\item{via}{}
}
\value{
data.frame with closest distance value for each input feature
}
\description{
ee_closest_distance_to_val
}
