% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_ls8_mask.R
\name{ee_l8mask}
\alias{ee_l8mask}
\title{Title}
\usage{
ee_l8mask(image, collection = 1, radix = 2, type = "clouds&shadows")
}
\arguments{
\item{image}{}
}
\value{

}
\description{
Title
}
\examples{
TODO refactor radix, cloud and shadow bits...write method for collection and tidyee
}
