% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_rolling_statiistic.R
\name{ee_rolling_statistic2}
\alias{ee_rolling_statistic2}
\title{ee_rolling_statistic2}
\usage{
ee_rolling_statistic2(
  x = monthly_rainfall,
  stat = "sum",
  window,
  time_unit = "month",
  return_tidyee = T
)
}
\arguments{
\item{x}{tidyee object}

\item{stat}{\code{character} rolling statistic e.g mean (default), median, min, max, sd}

\item{window}{\code{numeric} window size for rolling statistic}

\item{time_unit}{\code{character} time unit for statistic e.g "month" (default),"day"}
}
\value{
tidyee containing ee$ImageCollection where each image represents the right aligned rolling statistics specified
}
\description{
ee_rolling_statistic2
}
\examples{
\dontrun{
library(tidyrgee)
library(rgee)
library(dplyr)
ee_Initialize()
chirps_link <- "UCSB-CHG/CHIRPS/DAILY"
chirps <- ee$ImageCollection(chirps_link)
chirps_tidy <- as_tidyee(chirps)
rolling_tidyee_collection <- ee_rolling_statistic(x=chirps_tidy,stat="mean", window=3,time_unit="month")
}
}
