% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.R
\name{ee_spi}
\alias{ee_spi}
\title{ee_spi calculate standard precipitation index}
\usage{
ee_spi(
  x,
  window = 6,
  window_unit = "month",
  ic_time_unit = "month",
  band = "precipitation_sum",
  moi = c(3, 4),
  use_tidyee = T
)
}
\arguments{
\item{x}{tidyee imageCollection}

\item{window}{window of time to calculate sPI}

\item{window_unit}{unit o ftime to calculate}

\item{ic_time_unit}{imagecollection time unit}

\item{band}{name of band}

\item{moi}{month of interest. (i.e if moi =4 and window is 3 months then the SPI is FMA)}

\item{use_tidyee}{default = T , no option for F at the moment.}
}
\value{

}
\description{
ee_spi calculate standard precipitation index
}
\details{
By harnessing tidyee with `use_tidyee=T` we don't have any
}
\examples{
\dontrun{
library(rgee)
library(tidyrgee)
ee_Initialize()
spi_3 <- ee_spi(x=monthly_rainfall,
               window=3,
               window_unit="month",
               ic_time_unit= "month",
               band="precipitation_sum",
               moi=4,use_tidyee=T)
}
}
