% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_chirps_rain_intensity.R
\name{extract_chirps_rain_intensity}
\alias{extract_chirps_rain_intensity}
\title{extract_chirps_rain_intensity}
\usage{
extract_chirps_rain_intensity(geom_sf, from_when)
}
\arguments{
\item{geom_sf}{geometric unit to extract to}
}
\value{
data.frame with rainfall intensity for each unit of geom_sf
}
\description{
wrapper for ee_chirps_rain_intensity for purpose of creating target
}
