% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_growing_season_length_viirs.R
\name{extract_growing_season_length_viirs}
\alias{extract_growing_season_length_viirs}
\title{Title}
\usage{
extract_growing_season_length_viirs(geom_sf, yoi = 2013:2020, scale = 500)
}
\arguments{
\item{scale}{}
}
\value{

}
\description{
Title
}
