% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_modis_ndvi_anomaly.R
\name{extract_modis_ndvi_anomaly}
\alias{extract_modis_ndvi_anomaly}
\title{extract_ndvi_anomaly}
\usage{
extract_modis_ndvi_anomaly(
  geom_sf,
  baseline_years = 2000:2021,
  date_range = c("2021-06-20", "2021-09-26"),
  range_label = "growing_season",
  scale = 250
)
}
\arguments{
\item{scale}{}
}
\value{

}
\description{
extract_ndvi_anomaly
}
