% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nearest_water_pixel_distance.R
\name{extract_nearest_water_pixel_distance}
\alias{extract_nearest_water_pixel_distance}
\title{Title}
\usage{
extract_nearest_water_pixel_distance(
  y,
  water_type = "permanent",
  scale = 30,
  tidy_extract = T,
  year_range = c(2017, 2020),
  via
)
}
\arguments{
\item{tidy_extract}{}
}
\value{

}
\description{
Title
}
