% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_air_quality_idicators.R
\name{extract_s5p_air_quality}
\alias{extract_s5p_air_quality}
\title{Title}
\usage{
extract_s5p_air_quality(geom_sf, yoi = 2022, moi = 5, img_scale = 111320)
}
\arguments{
\item{img_scale}{}
}
\value{

}
\description{
Title
}
