% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_msna.R
\name{fetch_col_msna}
\alias{fetch_col_msna}
\title{Title}
\usage{
fetch_col_msna(country_code = "col")
}
\arguments{
\item{country_code}{}
}
\value{

}
\description{
redundant function I had to make when COL sent new data - did not want to edit `fetch_msna` as that
would caue me to have to create >90 targets ~5-10 hours... instead just make a new silly function. Probably using a method to monitor changes to input file could avoide this
}
