% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_msna.R
\name{fetch_msna_path}
\alias{fetch_msna_path}
\title{fetch_msna_path - helper function to quickly pull msna coordinate file path}
\usage{
fetch_msna_path(country_code = "col")
}
\arguments{
\item{country_code}{\code{character} three letter country code (lower-case)}
}
\value{
sf data.frame containing msna records
}
\description{
return msna data based on suplied country code
}
\examples{
\dontrun{
library(surveyGEER)
df <-  fetch_msna("com")
}
}
