% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_wrangling_functions.R
\name{format_rs_indicators_long}
\alias{format_rs_indicators_long}
\title{Title}
\usage{
format_rs_indicators_long(country_code, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
this output is more useful for my shiny or rmd reporting
}
