% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_wrangling_functions.R
\name{format_rs_indicators_wide}
\alias{format_rs_indicators_wide}
\title{Title}
\usage{
format_rs_indicators_wide(rs_long)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
 rs_indicators_long <- format_rs_indicators_long(col_pt_data_clean,
                                               col_chirps_rainfall_intensity,
                                                col_prev_3mo_drought_modis_basea,
                                                col_chirps_spi,
                                                col_dist_to_coast,
                                                col_landforms_reclassified,
                                                col_oxford_access,
                                                col_npp,
                                                col_local_value_merged)

rs_indicators_wide <- format_rs_indicators_wide(rs_indicators_long)
}
}
