% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geomorph_landform_indicators.R
\name{get_geomorph_landform_indicators}
\alias{get_geomorph_landform_indicators}
\title{get_geomorph_landforms}
\usage{
get_geomorph_landform_indicators(return_tidyee = F)
}
\value{
image or tidyee with image
}
\description{
get_geomorph_landforms
}
