% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modis_drought_indicators.R
\name{get_modis_drought_indicators}
\alias{get_modis_drought_indicators}
\title{get_modis_drought_indicators}
\usage{
get_modis_drought_indicators(
  baseline_years = c(2000:2015),
  yoi = c(2021:2022),
  mask = "none",
  satellite = "terra",
  TAC = F,
  temporal_interpolation = F
)
}
\arguments{
\item{yoi}{}
}
\value{
tidyee containing imageCollection with drought indicators
}
\description{
get_modis_drought_indicators
}
\examples{
so far a deliberate decision to not gap fill historical baseline.
This is due to the fact that by composting over x years we are already gap filling.
An improvement could potentially be to apply a conditional gap filling based on
\% masked pixels in each pixel-stack. If crosses the threshold could gap fill.
}
