% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_rolling_statiistic.R
\name{get_roll_dates}
\alias{get_roll_dates}
\title{get_roll_dates}
\usage{
get_roll_dates(x, window, time_unit)
}
\arguments{
\item{x}{tidyee object}

\item{window}{\code{numeric} time window to calculate rolling statistic on. Currently can only roll right}

\item{time_unit}{\code{character} unit of temporal aggregation (default = month).}
}
\value{
`ee$List` composed of `ee$Date` ready for using with ee$List$map
}
\description{
get_roll_dates
}
\examples{
\dontrun{
library(tidyrgee)
library(rgee)
library(dplyr)
ee_Initialize()
chirps_link <- "UCSB-CHG/CHIRPS/DAILY"
chirps <- ee$ImageCollection(chirps_link)
chirps_tidy <- as_tidyee(chirps)
dates_to_map <- get_roll_dates(x=chirps_tidy, window=3,time_unit="month")
}
}
