% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_ics.R
\name{interpolate_ic}
\alias{interpolate_ic}
\title{interpolate_ic}
\usage{
interpolate_ic(x, time_window)
}
\arguments{
\item{x}{ee$ImageCollection}

\item{time_window}{time range to look forward and backward for unmasked pixels}
}
\value{
ee$ImageCollection with bands interpolated
}
\description{
interpolate_ic
}
\details{
code from Ujaval Gandhi,2021
}
