% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_functions.R
\name{load_clean_assessement_points}
\alias{load_clean_assessement_points}
\title{Title}
\usage{
load_clean_assessement_points(country_code)
}
\arguments{
\item{country_code}{}
}
\value{

}
\description{
Title
}
