% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_functions.R
\name{load_clean_col_assessement_points}
\alias{load_clean_col_assessement_points}
\title{Title}
\usage{
load_clean_col_assessement_points(country_code)
}
\value{

}
\description{
redundant function I had to make when COL sent new data - did not want to edit `fetch_msna` as that
would caue me to have to create >90 targets ~5-10 hours... instead just make a new silly function. Probably using a method to monitor changes to input file could avoide this
}
