% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_report_functions.R
\name{make_overview_hex_map}
\alias{make_overview_hex_map}
\title{Title}
\usage{
make_overview_hex_map(sf_geom, country_code, hex_size)
}
\arguments{
\item{hex_size}{}
}
\value{

}
\description{
Title
}
