% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_date.R
\name{median_survey_date}
\alias{median_survey_date}
\title{calculate median date}
\usage{
median_survey_date(df, survey_date)
}
\arguments{
\item{survey_date}{column names of survey date}

\item{data.frame}{data frame}
}
\value{
median survey date
}
\description{
calculate median date
}
\examples{
\dontrun{
dat<- read_csv(here::here("assessment_data.csv"))
assessment_date <- median_survey_date(df = dat,survey_date = "survey_date")

}
}
