% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouple_coordinates.R
\name{merge_indicators2}
\alias{merge_indicators2}
\title{merge_indicators2}
\usage{
merge_indicators2(df, country_code = "col", df_uuid, hq = F)
}
\arguments{
\item{country_code}{\code{character} length 3 character with country code}

\item{hq}{\code{logical}  little argument i had to add after the fact in the case that 1 HQ analyst has to perform the workflow step for multiple teams at once}
}
\value{

}
\description{
merge_indicators2
}
