% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_modis_ndvi_anomaly.R
\name{modis_ndvi_anomaly}
\alias{modis_ndvi_anomaly}
\title{modis ndvi anomaly time range}
\usage{
modis_ndvi_anomaly(
  baseline_years = 2000:2021,
  date_range = c("2021-06-20", "2021-09-26"),
  range_label = "growing_season"
)
}
\arguments{
\item{date_range}{}
}
\value{

}
\description{
modis ndvi anomaly time range
}
