% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_wrangling_functions.R
\name{prep_rs_modis_target}
\alias{prep_rs_modis_target}
\title{Title}
\usage{
prep_rs_modis_target(df, uid_col = "new_uid")
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
library(targets)
tar_load_everything()
bla <- dplyr::lst(col_pt_data_clean,
    col_chirps_rainfall_intensity,
    col_prev_3mo_drought_modis_basea,
    col_chirps_spi,
    col_dist_to_coast,
    col_landforms_reclassified,
    col_oxford_access,
    col_npp,
    col_local_value_merged
    )

#'
}


}
