% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_reproject.R
\name{reach_epsg}
\alias{reach_epsg}
\title{reach_epsg}
\usage{
reach_epsg(country_code)
}
\arguments{
\item{country_code}{}
}
\value{
\code{numeric} epsg code
}
\description{
reach_epsg
}
\examples{
\dontrun{
library(surveyGEER)
reach_epsg("bgd")
}
}
