% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_reproject.R
\name{reach_reproject_utm}
\alias{reach_reproject_utm}
\title{reach_reproject_utm}
\usage{
reach_reproject_utm(x, country_code)
}
\arguments{
\item{x}{sf or sfc object}

\item{country_code}{lower case country code (i.e "bgd","som")}
}
\value{
x reprojected to specific WGS 84 UTM Zone
}
\description{
reach_reproject_utm
}
\examples{
\dontrun{
library(surveyGEER)
som_boundary <- surveyGEER::som_boundary
som_boundary |>
    reach_reproject_utm(country_code = "som")
}
}
