% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_air_quality_idicators.R
\name{s5p_air_quality}
\alias{s5p_air_quality}
\title{sentinel 5p air quality indicators}
\usage{
s5p_air_quality()
}
\value{
tidyee object `ee$ImageCollection` and S02 and No2 bands
}
\description{
sentinel 5p air quality indicators
}
\examples{
\dontrun{
library(rgee)
library(tidyrgee)
ee_Initialize()
daily_air_quality<-s5p_air_quality()
}
}
