% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_air_quality_idicators.R
\name{s5p_so2_ic}
\alias{s5p_so2_ic}
\title{Title}
\usage{
s5p_so2_ic(cloud_fraction_thresh = 0.05)
}
\arguments{
\item{cloud_fraction_thresh}{}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
so2_daily <- s5p_so2_ic()
}
}
