% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi_bar_plot.R
\name{spi_bar_plot_faceted}
\alias{spi_bar_plot_faceted}
\title{spi_bar_plot_faceted}
\usage{
spi_bar_plot_faceted(df, zone)
}
\arguments{
\item{df}{data.frame}

\item{zone}{zone column to facet by}
}
\value{
typical SPI barplot faceted by zone of choice
}
\description{
spi_bar_plot_faceted
}
\examples{
\dontrun{
library(surveyGEER)
library(rgee)
library(tidyrgee)
library(sf)
ee_Initialize() # initialize gee

#calculate spi
spi_3<- ee_chirps_spi(window = 3,window_unit = "month",moi = 4)
run zonal stat
spi_3_adm1 <- spi_3 |>
 ee_extract_tidy(y= som_adm1,
                 stat = "median",
                 scale=5500,
                 via="drive")
 # create bar plot
spi_bar_plot_faceted(spi_3_adm1,zone ="admin1name")
}

}
