% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_closest_distance_to_val.R
\name{switch_boolean}
\alias{switch_boolean}
\title{Title}
\usage{
switch_boolean(boolean, val)
}
\arguments{
\item{val}{}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
library(surveyGEER)
ee_bool<- switch_boolean(boolean = ">",val = 25)
}
}
