% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_modis_ndvi_anomaly.R
\name{terra_aqua_merge}
\alias{terra_aqua_merge}
\title{Title}
\usage{
terra_aqua_merge(tidyee = F)
}
\arguments{
\item{tidyee}{\code{logical} return tidyee class or image/imageCollection}
}
\value{

}
\description{
Title
}
