% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_img_stats_from_pt_sample.R
\name{zonal_stats_from_pt_sample}
\alias{zonal_stats_from_pt_sample}
\title{zonal stats from point sample}
\usage{
zonal_stats_from_pt_sample(x, aoi, size, stratified = F)
}
\arguments{
\item{x}{image/imageCollection}

\item{aoi}{sf or FC object representing area for extraction}

\item{size}{\code{numeric} number of sample}

\item{stratified}{\code{logical} does aoi contain zones to stratify sample by}
}
\value{
data.frame containing extracted values at points ... we should then
aggregate these to stats for the user.
}
\description{
zonal stats from point sample
}
